Submission 1
creates all the classes the HML diagram and the file that is being written to and read from, try catch and finally implemented successfully in each of the classes that need them to catch invalid inputs. The main is simple but effectively shows the functionality of each of the classes.